requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.6
lxml==4.9.4
requests==2.31.0
soupsieve==2.5
tls-client==0.2.2
urllib3==2.1.0
python-dotenv==1.0.0
